Towers Func (No enemy)
Towers are basically fully functional but only when shooting at the player not at enemies. Lots of debugging stuff was put in here, some of it not particularly useful (collision debugger scene and it's scripts), but the one useful one is being able to force the towers to shoot at the player rather than the enemy with space bar in the fight phase.